requirements.txt
# ═══════════════════════════════════════════════
# Parallax — Core Requirements
# ═══════════════════════════════════════════════

# LLM Providers
anthropic>=0.39.0
openai>=1.50.0
ollama>=0.3.0

# Embeddings & Vector Store
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Data Processing
pandas>=2.2.0
python-docx>=1.1.0
PyPDF2>=3.0.0
Pillow>=10.0.0
chardet>=5.2.0

# API & Platform
fastapi>=0.115.0
uvicorn>=0.30.0
websockets>=13.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Messaging Connectors (optional — install if using Slack/Discord)
# slack-sdk>=3.33.0
# discord.py>=2.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.8.0
loguru>=0.7.0
jinja2>=3.1.0
httpx>=0.27.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Data Export Parsers
beautifulsoup4>=4.12.0
lxml>=5.3.0
